+++ sub.go
@@ -2,22 +2,6 @@ package pubsub
 
 import "time"
 
-type PublishedMessage struct {
-	Topic string
-	Data  []byte
-}
-
-type SubscriptionRequest struct {
-	Subscriber   *Subscriber
-	ResponseChan chan uint64
-}
-
-type UnsubscriptionRequest struct {
-	Id           uint64
-	Topics       []string
-	ResponseChan chan uint64
-}
-
 type Subscriber struct {
 	Id      uint64
 	Channel chan *PublishedMessage
